% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.R
\name{BIC}
\alias{BIC}
\title{Title The Bayesian information criterion for selection of the tuning parameter}
\usage{
BIC(X, Status, Time, id, beta, stad, threshold)
}
\arguments{
\item{X}{a matrix of covariates that that may have effect on the failure times.}

\item{Status}{the censoring indicator, 1 = event of interest happens, and 0 = censoring.}

\item{Time}{right censored data which is the follow up time.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{beta}{the estimation of covariates coefficients.}

\item{stad}{if it is TRUE, all the covariates in the \code{formula} are standardized.}

\item{threshold}{estimated coefficients below \code{threshold} are considered as zero.}
}
\description{
Calculate the BIC score for selection of the tuning parameter.
}
