% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfun.R
\name{fitfun}
\alias{fitfun}
\title{Title  Selection of the tuning parameter for the penalized generalized estimating equations}
\usage{
fitfun(
  Time,
  Status,
  id,
  Var,
  X,
  corstr,
  stad,
  beta_int,
  nopindex = NULL,
  nlambda,
  lambda.min.ratio,
  eps,
  lambda = NULL,
  maxiter,
  tol,
  QIC
)
}
\arguments{
\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{Var}{if it is TRUE, the program returns Std.Error.}

\item{X}{a matrix of covariates that may have effect on the failure times.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}

\item{stad}{if it is TRUE, the covariates are standardized.}

\item{beta_int}{the initial value of the covariate coefficients.}

\item{nopindex}{a vector indicating which covariates are not penalized.}

\item{nlambda}{the number of \code{lambda} values.}

\item{lambda.min.ratio}{If the user does not provide \code{lambda}, the project will calculate the maximum of \code{lambda} and multiply it by \code{lambda.min.ratio} to determine the minimum of \code{lambda}. The default is \code{lambda.min.ratio=1e-4}.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{lambda}{a user supplied \code{lambda} sequence. Typical usage is to
have the program compute its own \code{lambda} sequence based on \code{nlambda}. Supplying a value of \code{lambda} overrides this.}

\item{maxiter}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{maxiter} iterations and the estimates will be based on the last iteration.}

\item{tol}{estimated coefficients below \code{tol} are considered as zero.}

\item{QIC}{if it is TRUE, the program returns the quasilikelihood information criterion score of the fitted model.}
}
\description{
We select the tuning parameters based on the BIC and fit the PTC model using SCAD-based penalized estimating equations.
}
