% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfun.R
\name{lambda_IS}
\alias{lambda_IS}
\title{calculate the maximum lambda value}
\usage{
lambda_IS(beta, Time, Status, X)
}
\arguments{
\item{beta}{the estimation of covariate coefficients.}

\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates that that may have effect on the failure times.}
}
\description{
calculate the maximum lambda value.
}
