% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfun.R
\name{Loss}
\alias{Loss}
\title{Loss function for beta}
\usage{
Loss(beta, Time, Status, X)
}
\arguments{
\item{beta}{the estimation of covariate coefficients.}

\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates that that may have effect on the failure times.}
}
\description{
Loss function for the promotion cure model.
}
