% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgeeptc.R
\name{print.pgeeptc}
\alias{print.pgeeptc}
\title{Title Print pgeeptc object}
\usage{
\method{print}{pgeeptc}(x, ...)
}
\arguments{
\item{x}{an object of \code{pgeeptc}.}

\item{...}{further arguments to be added in the \code{print.pgeeptc} function.}
}
\description{
Title Print pgeeptc object
}
