% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_H_E_M.R
\name{q_scad}
\alias{q_scad}
\title{Title The nonconvex SCAD penalty}
\usage{
q_scad(beta, lambda, a = 3.7)
}
\arguments{
\item{beta}{the estimation of covariates coefficients.}

\item{lambda}{the specified tuning parameter}

\item{a}{a parameter in the SCAD penalty. The default is \code{a=3.7}.}
}
\description{
Generate the SCAD penalty vector
}
