% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.R
\name{boots}
\alias{boots}
\title{Title Variance estimation based on the bootstrap method}
\usage{
boots(formula, data, id, nboot, corstr, beta_int, nopindex, lam, eps, maxiter)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}. The \code{response} is a \code{Surv} object with right censoring. The expression to the right of the "~" specifies the effect of covariates on the failure time.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula}.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{nboot}{the number of bootstrap samples.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}

\item{beta_int}{the initial value of the covariate coefficients.}

\item{nopindex}{a vector indicating which covariates are not penalized.}

\item{lam}{the tunning parameter in SCAD penalty.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{maxiter}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{maxiter} iterations and the estimates will be based on the last iteration. The default \code{maxiter = 100}.}
}
\description{
The variance estimation of the covariate coefficients and correlation coefficient in the working correlation matrix.
}
