% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycor.R
\name{mycor}
\alias{mycor}
\title{Title Estimation of the working correlation matrix and scale parameter}
\usage{
mycor(Status, id, X, beta_new, Time, corstr)
}
\arguments{
\item{Status}{the censoring indicator, 1 = event of interest happens, and 0 = censoring.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{X}{a matrix of covariates that that may have effect on the failure times.}

\item{beta_new}{the estimation of covariates coefficients.}

\item{Time}{right censored data which is the follow up time.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}
}
\description{
Estimate the working correlation matrix and scale parameter, where the structure of working correlation matrix can be specified as \code{independence}, \code{exchangeable} and \code{AR1}.
}
