% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgee.R
\name{pgee}
\alias{pgee}
\title{Title  The penalized generalized estimating equations}
\usage{
pgee(Time, Status, id, X, corstr, beta_int, nopindex, lam, eps, maxiter)
}
\arguments{
\item{Time}{right censored data which is the follow up time.}

\item{Status}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{X}{a matrix of covariates that may have effect on the failure times.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}.}

\item{beta_int}{the initial value of the covariate coefficients.}

\item{nopindex}{a vector indicating which covariates are not penalized.}

\item{lam}{the tunning parameter in SCAD penalty.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}.}

\item{maxiter}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{maxiter} iterations and the estimates will be based on the last iteration.}
}
\description{
The SCAD-based penalized estimating equations with the specified tuning parameter.
}
