% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgeeptc.R
\name{pgeeptc}
\alias{pgeeptc}
\title{Title  Variable selection in the marginal semiparametric promotion time cure model based on PGEE}
\usage{
pgeeptc(
  formula,
  id,
  data,
  corstr = "independence",
  stad = TRUE,
  Ibeta = NULL,
  Var = FALSE,
  lambda = NULL,
  nopindex = NULL,
  boots = FALSE,
  nboot = 100,
  QIC = FALSE,
  nlambda = 100,
  lambda.min.ratio = 1e-04,
  eps = 1e-05,
  maxiter = 100,
  tol = 0.001
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}. The \code{response} is a \code{Surv} object with right censoring. The expression to the right of the "~" specifies the effect of covariates on the failure time.}

\item{id}{a vector which identifies the clusters. The length of \code{id} should be the same as the number of observations.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula}.}

\item{corstr}{a character string specifying the correlation structure. The following are permitted: \code{independence}, \code{exchangeable} and \code{AR1}. The default is \code{corstr = "independence"}.}

\item{stad}{if it is TRUE, all the covariates in the \code{formula} are standardized. By default, \code{stdz = TRUE}.}

\item{Ibeta}{the initial value of the covariate coefficients.}

\item{Var}{if it is TRUE, the program returns Std.Error. By default, \code{Var = TRUE}.}

\item{lambda}{a user supplied \code{lambda} sequence. Typical usage is to
have the program compute its own \code{lambda} sequence based on \code{nlambda}. Supplying a value of \code{lambda} overrides this.}

\item{nopindex}{a vector indicating which covariates are not penalized.}

\item{boots}{if it is TRUE, the program returns Std.Error by the bootstrap method. By default, \code{boots = FALSE}.}

\item{nboot}{the number of bootstrap samples. The default is \code{nboot = 100}.}

\item{QIC}{if it is TRUE, the program returns the quasilikelihood information criterion score of the fitted model. By default, \code{boots = FALSE}.}

\item{nlambda}{the number of \code{lambda} values. The default is \code{nlambda = 100}.}

\item{lambda.min.ratio}{If the user does not provide \code{lambda}, the project will calculate the maximum of \code{lambda} and multiply it by \code{lambda.min.ratio} to determine the minimum of \code{lambda}. The default is \code{lambda.min.ratio=1e-4}.}

\item{eps}{tolerance for convergence. Iteration stops once the relative change in deviance is less than \code{eps}. The default is \code{eps = 1e-5}.}

\item{maxiter}{specifies the maximum iteration number. If the convergence criterion is not met, the iteration will be stopped after \code{maxiter} iterations and the estimates will be based on the last iteration. The default \code{maxiter = 100}.}

\item{tol}{estimated coefficients below \code{tol} are considered as zero. The default is \code{tol = 1e-3}.}
}
\value{
An object of class \code{pgeeptc} is returned. It can be examined by \code{print.pgeeptc()}.
}
\description{
Variable selection in the marginal semiparametric promotion time cure  (PTC) model based on penalized general estimation equations (PGEE) method.
We consider three common correlation structures in this funciton.
The variances can be consistently estimated by a sandwich variance estimator. We also present a bootstrap procedure for variance estimation.
}
