\name{print.pgeeptc}
\alias{print.pgeeptc}
\title{Print pgeeptc object}
\usage{
print.pgeeptc(x, ...)
}
\arguments{
\item{x}{an object of \code{pgeeptc}.}
\item{...}{further arguments to be added in the \code{print.pgeeptc} function.}
}
\description{
Print the summary of estimation and other details.
}
